fix: rename test to tst
